Only use one global graph and add v_net to wrapper attribute
Instead of a graph per scope, one global graph is used.
This is done because igraph has no subgraph support in c and
union functions can not handle attributes. Further, the id
shifts wuld be a pain to handle.

The wrapper gets a v_net filed as in the attribute structure which
can be used to connect it correctly in the future